implement atf_bagfile_plotter -- probably makes analysing step in atf_test obsolete for our purposes set speed_factor_analysis to 1.0!!! add some docu to config files take atf_presenter presenter.py from srd-pz -- firefox.sh did not show result for path_smoothness -> see 	readme contribution section for fix add metrics orientation_variation_/velocity_/acceleration_average, _max, _median add arg whether to start rviz some minor changes

+++ ipa_navigation_planning_atf/src/robot_route_recorder.cpp
@@ -57,7 +57,7 @@ void RobotRouteRecorder::makeFileFromData()
   file << "        \"orientation_z\": " + std::to_string(start_pose_.pose.orientation.z) +"\n";
   file << "        \"orientation_w\": " + std::to_string(start_pose_.pose.orientation.w) +"\n";
   file << "    \"/atf_test/goals\":\n";
-  for (int i = 0; i < goals_.size(); i++)
+  for (size_t i = 0; i < goals_.size(); i++)
   {
     file << "        - \"position_x\": " + std::to_string(goals_[i].position.x) +"\n";
     file << "          \"position_y\": " + std::to_string(goals_[i].position.y) +"\n";
